Moves api client to storage
